created UserData type
