+++ frontend/src/types/UserData.ts
@@ -1,0 +1,11 @@
+export interface UserData {
+    user: {
+        nickname: string
+        avatarUrl: string
+        nbVictory: number
+        totalPlay: number
+        xp: number
+        TFAEnabled: boolean
+        userPosition: number
+    }
+}
